+++ tests/exe.test.ts
@@ -1,31 +1,44 @@
 // run "EXE Mocha TS File - Loki" in VS Code to run this test
-import * as assert from 'assert';
-import * as Azure from 'azure-storage';
-import { execFile } from 'child_process';
-import find from 'find-process';
+import * as assert from "assert";
+import * as Azure from "azure-storage";
+import { execFile } from "child_process";
+import find from "find-process";
 
 import {
-    BlobServiceClient, newPipeline as blobNewPipeline,
-    StorageSharedKeyCredential as blobStorageSharedKeyCredential
-} from '@azure/storage-blob';
+  BlobServiceClient,
+  newPipeline as blobNewPipeline,
+  StorageSharedKeyCredential as blobStorageSharedKeyCredential
+} from "@azure/storage-blob";
 import {
-    newPipeline as queueNewPipeline, QueueClient, QueueServiceClient,
-    StorageSharedKeyCredential as queueStorageSharedKeyCredential
-} from '@azure/storage-queue';
+  newPipeline as queueNewPipeline,
+  QueueClient,
+  QueueServiceClient,
+  StorageSharedKeyCredential as queueStorageSharedKeyCredential
+} from "@azure/storage-queue";
 
-import { configLogger } from '../src/common/Logger';
-import { HeaderConstants, TABLE_API_VERSION } from '../src/table/utils/constants';
-import BlobTestServerFactory from './BlobTestServerFactory';
+import { configLogger } from "../src/common/Logger";
 import {
-    createConnectionStringForTest, HOST, PORT, PROTOCOL
-} from './table/apis/table.entity.test.utils';
+  HeaderConstants,
+  TABLE_API_VERSION
+} from "../src/table/utils/constants";
+import BlobTestServerFactory from "./BlobTestServerFactory";
 import {
-    bodyToString, EMULATOR_ACCOUNT_KEY, EMULATOR_ACCOUNT_NAME, getUniqueName, overrideRequest,
-    restoreBuildRequestOptions
-} from './testutils';
+  createConnectionStringForTest,
+  HOST,
+  PORT,
+  PROTOCOL
+} from "./table/utils/table.entity.test.utils";
+import {
+  bodyToString,
+  EMULATOR_ACCOUNT_KEY,
+  EMULATOR_ACCOUNT_NAME,
+  getUniqueName,
+  overrideRequest,
+  restoreBuildRequestOptions
+} from "./testutils";
 
-// server address used for testing. Note that Azurite.exe has 
-// server address of http://127.0.0.1:10000 and so on by default 
+// server address used for testing. Note that Azurite.exe has
+// server address of http://127.0.0.1:10000 and so on by default
 // and we need to configure them when starting azurite.exe
 const blobAddress = "http://127.0.0.1:11000";
 const queueAddress = "http://127.0.0.1:11001";
@@ -41,7 +54,6 @@ configLogger(false);
 const testLocalAzuriteInstance = true;
 
 describe("exe test", () => {
-  
   const tableService = Azure.createTableService(
     createConnectionStringForTest(testLocalAzuriteInstance)
   );
@@ -51,24 +63,39 @@ describe("exe test", () => {
 
   const requestOverride = { headers: {} };
 
-  let childPid : number;
+  let childPid: number;
 
   before(async () => {
     overrideRequest(requestOverride, tableService);
     tableName = getUniqueName("table");
-    const child = execFile(".\\release\\azurite.exe", ["--blobPort 11000", "--queuePort 11001", "--tablePort 11002"], {cwd: process.cwd(), shell: true, env: {}});
+    const child = execFile(
+      ".\\release\\azurite.exe",
+      ["--blobPort 11000", "--queuePort 11001", "--tablePort 11002"],
+      { cwd: process.cwd(), shell: true, env: {} }
+    );
 
     childPid = child.pid;
 
-    const fullSuccessMessage = "Azurite Blob service is starting at " + blobAddress + "\nAzurite Blob service is successfully listening at " + blobAddress + 
-                               "\nAzurite Queue service is starting at " + queueAddress + "\nAzurite Queue service is successfully listening at " + queueAddress + 
-                               "\nAzurite Table service is starting at " + tableAddress + "\nAzurite Table service is successfully listening at " + tableAddress + "\n";
-    let messageReceived : string = "";
+    const fullSuccessMessage =
+      "Azurite Blob service is starting at " +
+      blobAddress +
+      "\nAzurite Blob service is successfully listening at " +
+      blobAddress +
+      "\nAzurite Queue service is starting at " +
+      queueAddress +
+      "\nAzurite Queue service is successfully listening at " +
+      queueAddress +
+      "\nAzurite Table service is starting at " +
+      tableAddress +
+      "\nAzurite Table service is successfully listening at " +
+      tableAddress +
+      "\n";
+    let messageReceived: string = "";
 
     function stdoutOn() {
-      return new Promise(resolve => {
+      return new Promise((resolve) => {
         // exclamation mark suppresses the TS error that "child.stdout is possibly null"
-        child.stdout!.on('data', function(data: any) {
+        child.stdout!.on("data", function (data: any) {
           messageReceived += data.toString();
           if (messageReceived == fullSuccessMessage) {
             resolve("resolveMessage");
@@ -83,15 +110,15 @@ describe("exe test", () => {
   after(async () => {
     // TO DO
     // Currently, the mocha test does not quit unless "--exit" is added to the mocha command
-    // The current fix is to have "--exit" added but the issue causing mocha to be unable to 
+    // The current fix is to have "--exit" added but the issue causing mocha to be unable to
     // quit has not been identified
     restoreBuildRequestOptions(tableService);
     tableService.removeAllListeners();
 
-    await find('name', 'azurite.exe', true).then((list: any) => {
+    await find("name", "azurite.exe", true).then((list: any) => {
       process.kill(list[0].pid);
     });
-    
+
     process.kill(childPid);
   });
 
@@ -107,7 +134,7 @@ describe("exe test", () => {
         Prefer: "return-no-content",
         accept: "application/json;odata=minimalmetadata"
       };
-  
+
       tableService.createTable(tableName, (error, result, response) => {
         if (!error) {
           assert.strictEqual(result.TableName, tableName);
@@ -119,7 +146,7 @@ describe("exe test", () => {
         done();
       });
     });
-  
+
     it("queryTable, accept=application/json;odata=minimalmetadata @loki", (done) => {
       /* Azure Storage Table SDK doesn't support customize Accept header and Prefer header,
         thus we workaround this by override request headers to test following 3 OData levels responses.
@@ -130,55 +157,61 @@ describe("exe test", () => {
       requestOverride.headers = {
         accept: "application/json;odata=minimalmetadata"
       };
-  
-      tableService.listTablesSegmented(null as any, (error, result, response) => {
-        if (!error) {
-          assert.strictEqual(response.statusCode, 200);
-          const headers = response.headers!;
-          assert.strictEqual(headers["x-ms-version"], TABLE_API_VERSION);
-          const bodies = response.body! as any;
-          assert.deepStrictEqual(
-            bodies["odata.metadata"],
-            `${PROTOCOL}://${HOST}:${PORT}/${EMULATOR_ACCOUNT_NAME}/$metadata#Tables`
-          );
-          assert.ok(bodies.value[0].TableName);
+
+      tableService.listTablesSegmented(
+        null as any,
+        (error, result, response) => {
+          if (!error) {
+            assert.strictEqual(response.statusCode, 200);
+            const headers = response.headers!;
+            assert.strictEqual(headers["x-ms-version"], TABLE_API_VERSION);
+            const bodies = response.body! as any;
+            assert.deepStrictEqual(
+              bodies["odata.metadata"],
+              `${PROTOCOL}://${HOST}:${PORT}/${EMULATOR_ACCOUNT_NAME}/$metadata#Tables`
+            );
+            assert.ok(bodies.value[0].TableName);
+          }
+          done();
         }
-        done();
-      });
+      );
     });
-    
+
     it("deleteTable that exists, @loki", (done) => {
       /*
       https://docs.microsoft.com/en-us/rest/api/storageservices/delete-table
       */
       requestOverride.headers = {};
-  
+
       const tableToDelete = tableName + "del";
-  
+
       tableService.createTable(tableToDelete, (error, result, response) => {
         if (!error) {
-          tableService.deleteTable(tableToDelete, (deleteError, deleteResult) => {
-            if (!deleteError) {
-              // no body expected, we expect 204 no content on successful deletion
-              assert.strictEqual(deleteResult.statusCode, 204);
-            } else {
-              assert.ifError(deleteError);
+          tableService.deleteTable(
+            tableToDelete,
+            (deleteError, deleteResult) => {
+              if (!deleteError) {
+                // no body expected, we expect 204 no content on successful deletion
+                assert.strictEqual(deleteResult.statusCode, 204);
+              } else {
+                assert.ifError(deleteError);
+              }
+              done();
             }
-            done();
-          });
+          );
         } else {
           assert.fail("Test failed to create the table");
           done();
         }
       });
     });
-  
+
     it("deleteTable that does not exist, @loki", (done) => {
       // https://docs.microsoft.com/en-us/rest/api/storageservices/delete-table
       requestOverride.headers = {};
-  
+
       const tableToDelete = tableName + "causeerror";
-  
+
       tableService.deleteTable(tableToDelete, (error, result) => {
         assert.strictEqual(result.statusCode, 404); // no body expected, we expect 404
         const storageError = error as any;
@@ -186,17 +219,17 @@ describe("exe test", () => {
         done();
       });
     });
-  
+
     it("createTable with invalid version, @loki", (done) => {
       requestOverride.headers = { [HeaderConstants.X_MS_VERSION]: "invalid" };
-  
+
       tableService.createTable("test", (error, result) => {
         assert.strictEqual(result.statusCode, 400);
         done();
       });
     });
   });
-  
+
   describe("blob test", () => {
     const factory = new BlobTestServerFactory();
     const blobServer = factory.createServer();
@@ -239,14 +272,17 @@ describe("exe test", () => {
     });
     it("download with with default parameters @loki @sql", async () => {
       const result = await blobClient.download(0);
-      assert.deepStrictEqual(await bodyToString(result, content.length), content);
+      assert.deepStrictEqual(
+        await bodyToString(result, content.length),
+        content
+      );
       assert.equal(result.contentRange, undefined);
       assert.equal(
         result._response.request.headers.get("x-ms-client-request-id"),
         result.clientRequestId
       );
     });
-  
+
     it("download should work with conditional headers @loki @sql", async () => {
       const properties = await blobClient.getProperties();
       const result = await blobClient.download(0, undefined, {
@@ -257,7 +293,10 @@ describe("exe test", () => {
           ifUnmodifiedSince: new Date("2188/01/01")
         }
       });
-      assert.deepStrictEqual(await bodyToString(result, content.length), content);
+      assert.deepStrictEqual(
+        await bodyToString(result, content.length),
+        content
+      );
       assert.equal(result.contentRange, undefined);
       assert.equal(
         result._response.request.headers.get("x-ms-client-request-id"),
@@ -309,7 +348,7 @@ describe("exe test", () => {
         mResult._response.request.headers.get("x-ms-client-request-id"),
         mResult.clientRequestId
       );
-  
+
       const result = await queueClient.getProperties();
       assert.deepEqual(result.metadata, metadata);
       assert.equal(
@@ -317,7 +356,7 @@ describe("exe test", () => {
         result.clientRequestId
       );
     });
-  
+
     it("getProperties with default/all parameters @loki", async () => {
       const result = await queueClient.getProperties();
       assert.ok(result.approximateMessagesCount! >= 0);
@@ -325,7 +364,7 @@ describe("exe test", () => {
       assert.ok(result.version);
       assert.ok(result.date);
     });
-  
+
     it("SetAccessPolicy should work @loki", async () => {
       const queueAcl = [
         {
@@ -345,13 +384,13 @@ describe("exe test", () => {
           id: "policy2"
         }
       ];
-  
+
       const sResult = await queueClient.setAccessPolicy(queueAcl);
       assert.equal(
         sResult._response.request.headers.get("x-ms-client-request-id"),
         sResult.clientRequestId
       );
-  
+
       const result = await queueClient.getAccessPolicy();
       assert.deepEqual(result.signedIdentifiers, queueAcl);
       assert.equal(
@@ -361,4 +400,3 @@ describe("exe test", () => {
     });
   });
 });
-